pyproject.toml
[project]
name = "gemini-llm-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=0.5.23",
    "crewai-tools>=0.47.1",
    "google-adk>=1.3.0",
    "google-genai>=1.20.0",
    "langchain>=0.3.25",
    "langchain-community>=0.3.25",
    "langchain-google-genai>=2.1.5",
    "pdf2image>=1.17.0",
    "pi-heif>=0.22.0",
    "pillow-heif>=0.22.0",
    "playwright>=1.52.0",
    "pypdf>=5.6.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
    "unstructured>=0.17.2",
    "unstructured-inference>=1.0.5",
    "unstructured-pytesseract>=0.3.15",
]
